dom/base/test/test_window_content.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1400139
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1400139</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1400139 **/
  var props = [];
  for (var prop in window) props.push(prop);
  is(props.indexOf("content"), -1, "Should not have a property named 'content'");
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1400139">Mozilla Bug 1400139</a>
<p id="display"></p>
<div id="notcontent" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
